regras de negocio implementadas

+++ VendaPro/Models/Produto.cs
@@ -5,18 +5,18 @@ namespace VendaPro.Models
         public int Id { get; set; }
         public string Nome { get; set; }
         public int Quantidade { get; set; }
-        public double Preco { get; set; }
+        public decimal Preco { get; set; }
 
         public Produto() { }
 
-        public Produto(string nome, int quantidade, double preco)
+        public Produto(string nome, int quantidade, decimal preco)
         {
             Nome = nome;
             Quantidade = quantidade;
             Preco = preco;
         }
 
-        public double CalcularDesconto(double percentual)
+        public decimal CalcularDesconto(decimal percentual)
         {
             if (percentual < 0 || percentual > 100)
                 throw new ArgumentException("Percentual inválido.");
